# state/runs/2026-10-09-040829-umm-improve/validated-snapshot/docs/openapi.yaml
openapi: 3.1.0
info:
  title: umm API
  version: 0.8.1
  description: |
    Spatial Thought Memory, collaboration, automation and Dream Layer API.
    Supported Canvas mutation requests can include an `Idempotency-Key` header (8-128 safe characters) for replay-safe offline synchronization.
    Personal preferences, API-key lifecycle operations and all administrator operations require an interactive browser session; bearer API keys cannot call them.
    Errors use RFC 9457 Problem Details and note version conflicts include the latest server note.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /handoff/claims:
    post:
      operationId: post_handoff_claims
      tags: [Thoughts]
      summary: Issue a single-use claim for a space, to hand it to another service
      description: |
        The sending half of the in-house handoff standard. The claim is bound
        to one space this person may read, is good for five minutes and for one
        collection, and widens nothing: the space is checked the way every
        export is, through the same approval gate, and the document is rendered
        as this person. The document is the space's outline — its sentences in
        the order the graph puts them — rendered now and kept with the claim,
        so what is collected is exactly what `bytes` announced.

        The claim never reaches the audit log or the access log.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource, format]
              properties:
                resource: { type: string, format: uuid, description: The space. }
                format: { type: string, enum: [markdown], description: umm sends Markdown and nothing else. }
      responses:
        '201':
          description: The claim
          content:
            application/json:
              schema:
                type: object
                required: [claim, source, filename, content_type, bytes, expires_at]
                properties:
                  claim: { type: string, description: 256 bits of randomness, URL-safe. }
                  source: { type: string, format: uri, description: Where to collect it from — this service's origin. }
                  filename: { type: string }
                  content_type: { type: string, enum: ['text/markdown; charset=utf-8'] }
                  bytes: { type: integer }
                  expires_at: { type: string, format: date-time }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: A team lead has to approve this export first
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  code: { type: string, enum: [approval_required] }
  /handoff/claims/{claim}:
    parameters:
      - { in: path, name: claim, required: true, schema: { type: string } }
    get:
      operationId: get_handoff_claims_claim
      security: []
      tags: [Thoughts]
      summary: Collect a handed-off document; the claim is the credential
      description: |
        No sign-in. The claim is spent by this request: a second collection,
        an expired claim and a claim that was never issued all answer 404, and
        the answer does not say which.
      responses:
        '200':
          description: The document, named by Content-Disposition
          headers:
            Content-Disposition: { schema: { type: string }, description: "attachment; filename=\"...\"; filename*=UTF-8''..." }
          content:
            text/markdown:
              schema: { type: string }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: umm_session }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: umm_key }
...
  responses:
...
    Error: { description: RFC 9457 Problem Details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, error]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        error: { type: string, description: Backward-compatible alias for detail }
        requestId: { type: string }
        clientVersion: { type: integer }
        latest: { $ref: '#/components/schemas/Note' }
...
    NoteWrite:
      type: object
      required: [content, x, y]
      properties:
        content: { type: string }
        title: { type: string }
        color: { type: string, enum: [yellow, blue, purple, green, red, gray, lavender] }
        kind:
          type: string
          enum: [thought, question, idea]
          default: thought
          description: |
            What the thought is. Marked, never inferred: umm does not read a note
            and decide it is a question. A value outside the vocabulary is
            rejected with 400 and an allowedKinds list rather than rewritten,
            which would file a thought as something the person did not say it was.
            "idea" is written when a Dream is materialised.
        aiExcluded: { type: boolean, description: Exclude this note from AI processing and remote embedding gateways }
        x: { type: number }
        y: { type: number }
        width: { type: number, minimum: 120 }
        height: { type: number, minimum: 80 }
        rotation: { type: number }
    Note:
      allOf:
        - { $ref: '#/components/schemas/NoteWrite' }
        - type: object
          required: [id, spaceId, authorId, source, version, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            spaceId: { type: string, format: uuid }
            authorId: { type: string, format: uuid }
            source: { type: string, enum: [user, dream, api, mcp] }
            version: { type: integer }
            relatedCount: { type: integer, minimum: 0 }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...